requirements.txt
opencv-python
numpy
Flask
playwright
cryptography
Pillow